Enhance object method invocation evaluation support
