+++ misc/debug-adapter/modules/debug-adapter-core/src/main/java/org/ballerinalang/debugadapter/evaluation/engine/GeneratedInstanceMethod.java
@@ -16,7 +16,9 @@
 
 package org.ballerinalang.debugadapter.evaluation.engine;
 
+import com.sun.jdi.AbsentInformationException;
 import com.sun.jdi.ClassNotLoadedException;
+import com.sun.jdi.LocalVariable;
 import com.sun.jdi.Method;
 import com.sun.jdi.ObjectReference;
 import com.sun.jdi.Type;
@@ -29,6 +31,9 @@ import org.ballerinalang.debugadapter.evaluation.EvaluationUtils;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+
+import static org.ballerinalang.debugadapter.evaluation.EvaluationUtils.STRAND_VAR_NAME;
 
 /**
  * JVM generated instance method representation of a ballerina function.
@@ -38,6 +43,7 @@ import java.util.Map;
 public class GeneratedInstanceMethod extends JvmMethod {
 
     private final Value objectValueRef;
+    protected Map<String, Value> namedArgValues;
 
     GeneratedInstanceMethod(SuspendedContext context, Value objectRef, Method methodRef) {
         super(context, methodRef);
@@ -48,6 +54,10 @@ public class GeneratedInstanceMethod extends JvmMethod {
                             List<Map.Entry<String, Evaluator>> argEvaluators, List<Value> argsList) {
         super(context, methodRef, argEvaluators, argsList);
         this.objectValueRef = objectRef;
+    }
+
+    public void setNamedArgValues(Map<String, Value> namedArgValues) {
+        this.namedArgValues = namedArgValues;
     }
 
     @Override
@@ -74,7 +84,7 @@ public class GeneratedInstanceMethod extends JvmMethod {
     @Override
     protected List<Value> getMethodArgs(JvmMethod method) throws EvaluationException {
         try {
-            if (argValues == null && argEvaluators == null) {
+            if (argValues == null && argEvaluators == null && namedArgValues == null) {
                 throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR
                         .getString(), methodRef.name()));
             }
@@ -89,6 +99,21 @@ public class GeneratedInstanceMethod extends JvmMethod {
                 // Here we use the existing strand instance to execute the function invocation expression.
                 Value strand = getCurrentStrand();
                 argValueList.add(0, strand);
+                return argValueList;
+            }
+
+            if (namedArgValues != null && !namedArgValues.isEmpty()) {
+                // Here we use the existing strand instance to execute the function invocation expression.
+                Value strand = getCurrentStrand();
+                namedArgValues.put(STRAND_VAR_NAME, strand);
+                List<LocalVariable> args = method.methodRef.arguments();
+                List<String> argNames = args.stream().map(LocalVariable::name).collect(Collectors.toList());
+                argNames.forEach(argName -> {
+                    argValueList.add(namedArgValues.get(argName));
+                    if (!argName.equals(STRAND_VAR_NAME)) {
+                        argValueList.add(EvaluationUtils.make(context, true).getJdiValue());
+                    }
+                });
                 return argValueList;
             }
 
@@ -110,7 +135,7 @@ public class GeneratedInstanceMethod extends JvmMethod {
             Value strand = getCurrentStrand();
             argValueList.add(0, strand);
             return argValueList;
-        } catch (ClassNotLoadedException e) {
+        } catch (ClassNotLoadedException | AbsentInformationException e) {
             throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR.getString(),
                     methodRef.name()));
         }
